1차 조회 api

+++ src/main/kotlin/com/hansol/recordsserver/common/constants/PageSize.kt
@@ -1,5 +1,5 @@
 package com.hansol.recordsserver.common.constants
 
 object PageSize {
-    const val DEFAULT_PAGE_SIZE: Int = 20
+    const val DEFAULT_PAGE_SIZE: Long = 20L
 }
